requirements.txt
torch
torchaudio
torchvision
numpy
pandas
scikit-learn
matplotlib
pyyaml
optuna
tqdm
tensorboard